requirements.txt
flask>=2.3.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-whois>=0.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-Levenshtein>=0.21.0
sqlalchemy>=2.0.0
pytest>=7.4.0
joblib>=1.3.0
gradio>=4.0.0
